pyproject.toml
[tool.poetry]
name = "django-tutorial"
version = "0.1.0"
description = ""
authors = ["leon <90852553+leon0719@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "django_tutorial"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
psycopg2 = "^2.9.9"
pillow = "^10.1.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
black = "^23.12.1"
pytest-django = "^4.8.0"
pytest-xdist = "^3.5.0"
faker = "^24.3.0"
pre-commit = "^3.6.2"
ruff = "^0.3.4"
pytest-cov = "^5.0.0"

[tool.ruff.lint]
extend-ignore = ["E401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
